[OpenAI] Prepare release
