+++ sdk/core/core-client-rest/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@azure-rest/core-client",
-  "version": "1.1.7",
+  "version": "1.2.0",
   "description": "Core library for interfacing with Azure Rest Clients",
   "sdk-type": "client",
   "main": "dist/index.js",
@@ -58,7 +58,7 @@
     "@azure/core-auth": "^1.3.0",
     "@azure/core-rest-pipeline": "^1.5.0",
     "@azure/core-util": "^1.0.0",
-    "@azure/abort-controller": "^1.1.0",
+    "@azure/abort-controller": "^2.0.0",
     "@azure/core-tracing": "^1.0.1",
     "tslib": "^2.2.0"
   },
