+++ sdk/openai/openai/samples/v1-beta/javascript/bringYourOwnData.js
@@ -34,7 +34,7 @@ async function main() {
 
   const client = new OpenAIClient(endpoint, new AzureKeyCredential(azureApiKey));
   const deploymentId = "gpt-35-turbo";
-  const events = client.listChatCompletions(deploymentId, messages, {
+  const events = await client.streamChatCompletions(deploymentId, messages, {
     maxTokens: 128,
     /**
      * The `azureExtensionOptions` property is used to configure the
