+++ sdk/openai/openai/samples/v1-beta/javascript/streamCompletions.js
@@ -7,11 +7,10 @@
  * @summary list completions.
  */
 
-import { OpenAIClient, AzureKeyCredential } from "@azure/openai";
+const { OpenAIClient, AzureKeyCredential } = require("@azure/openai");
 
 // Load the .env file if it exists
-import * as dotenv from "dotenv";
-dotenv.config();
+require("dotenv").config();
 
 // You will need to set these environment variables or edit the following values
 const endpoint = process.env["ENDPOINT"] || "<endpoint>";
@@ -19,12 +18,12 @@ const azureApiKey = process.env["AZURE_API_KEY"] || "<api key>";
 
 const prompt = ["What is Azure OpenAI?"];
 
-export async function main() {
+async function main() {
   console.log("== Stream Completions Sample ==");
 
   const client = new OpenAIClient(endpoint, new AzureKeyCredential(azureApiKey));
   const deploymentId = "text-davinci-003";
-  const events = client.listCompletions(deploymentId, prompt, { maxTokens: 128 });
+  const events = await client.streamCompletions(deploymentId, prompt, { maxTokens: 128 });
 
   for await (const event of events) {
     for (const choice of event.choices) {
@@ -36,3 +35,5 @@ export async function main() {
 main().catch((err) => {
   console.error("The sample encountered an error:", err);
 });
+
+module.exports = { main };
