+++ sdk/openai/openai/samples/v1-beta/typescript/README.md
@@ -13,20 +13,19 @@ urlFragment: openai-typescript-beta
 
 These sample programs show how to use the TypeScript client libraries for Azure OpenAI in some common scenarios.
 
-| **File Name**                                                                   | **Description**                                                          |
-| ------------------------------------------------------------------------------- | ------------------------------------------------------------------------ |
-| [audioTranscription.ts][audiotranscription]                                     | audio transcription.                                                     |
-| [audioTranslation.ts][audiotranslation]                                         | audio translation.                                                       |
-| [bringYourOwnData.ts][bringyourowndata]                                         | chat completions with your own data.                                     |
-| [chatCompletions.ts][chatcompletions]                                           | get chat completions.                                                    |
-| [completions.ts][completions]                                                   | get completions.                                                         |
-| [functions.ts][functions]                                                       | get chat completions with functions.                                     |
-| [getImages.ts][getimages]                                                       | generates images from prompts using Azure OpenAI Batch Image Generation. |
-| [listChatCompletions.ts][listchatcompletions]                                   | list chat completions.                                                   |
-| [listChatCompletionsWithContentFilter.ts][listchatcompletionswithcontentfilter] | get completions.                                                         |
-| [listCompletions.ts][listcompletions]                                           | list completions.                                                        |
-| [readableStream.ts][readablestream]                                             | stream chat completions.                                                 |
-| [openAi.ts][openai]                                                             | get completions using the OpenAI API.                                    |
+| **File Name**                                                                       | **Description**                                                          |
+| ----------------------------------------------------------------------------------- | ------------------------------------------------------------------------ |
+| [audioTranscription.ts][audiotranscription]                                         | audio transcription.                                                     |
+| [audioTranslation.ts][audiotranslation]                                             | audio translation.                                                       |
+| [bringYourOwnData.ts][bringyourowndata]                                             | chat completions with your own data.                                     |
+| [chatCompletions.ts][chatcompletions]                                               | get chat completions.                                                    |
+| [completions.ts][completions]                                                       | get completions.                                                         |
+| [functions.ts][functions]                                                           | get chat completions with functions.                                     |
+| [getImages.ts][getimages]                                                           | generates images from prompts using Azure OpenAI Batch Image Generation. |
+| [streamChatCompletions.ts][streamchatcompletions]                                   | list chat completions.                                                   |
+| [streamChatCompletionsWithContentFilter.ts][streamchatcompletionswithcontentfilter] | get completions.                                                         |
+| [streamCompletions.ts][streamcompletions]                                           | list completions.                                                        |
+| [openAi.ts][openai]                                                                 | get completions using the OpenAI API.                                    |
 
 ## Prerequisites
 
@@ -87,10 +86,9 @@ Take a look at our [API Documentation][apiref] for more information about the AP
 [completions]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/completions.ts
 [functions]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/functions.ts
 [getimages]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/getImages.ts
-[listchatcompletions]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/listChatCompletions.ts
-[listchatcompletionswithcontentfilter]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/listChatCompletionsWithContentFilter.ts
-[listcompletions]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/listCompletions.ts
-[readablestream]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/readableStream.ts
+[streamchatcompletions]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/streamChatCompletions.ts
+[streamchatcompletionswithcontentfilter]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/streamChatCompletionsWithContentFilter.ts
+[streamcompletions]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/streamCompletions.ts
 [openai]: https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/openai/openai/samples/v1-beta/typescript/src/openAi.ts
 [apiref]: https://docs.microsoft.com/javascript/api/@azure/openai
 [freesub]: https://azure.microsoft.com/free/
