+++ sdk/openai/openai/samples/v1-beta/typescript/src/bringYourOwnData.ts
@@ -35,7 +35,7 @@ export async function main() {
 
   const client = new OpenAIClient(endpoint, new AzureKeyCredential(azureApiKey));
   const deploymentId = "gpt-35-turbo";
-  const events = client.listChatCompletions(deploymentId, messages, {
+  const events = await client.streamChatCompletions(deploymentId, messages, {
     maxTokens: 128,
     /**
      * The `azureExtensionOptions` property is used to configure the
