+++ sdk/openai/openai/samples/v1-beta/typescript/src/functions.ts
@@ -46,7 +46,7 @@ export async function main() {
     [{ role: "user", content: "What's the weather like in Boston?" }],
     {
       functions: [getCurrentWeather],
-    }
+    },
   );
 
   for (const choice of result.choices) {
